final version, vectorized accessibility
