+++ spint/tests/test_accessibility.py
@@ -9,20 +9,17 @@ __author__ = 'Lenka Hasova haska.lenka@gmail.com'
 
 import unittest
 import numpy as np
-from ..generate_dummy_accessibility import generate_dummy_flows
+from ..flow_accessibility import generate_dummy_flows
 from ..flow_accessibility import Accessibility
-from ..flow_accessibility import AFED
+
 
 
 class AccessibilityTest(unittest.TestCase):
     
     def test_accessibility(self):
-        
         flow = generate_dummy_flows()
-        
         flow = flow.loc[:,['origin_ID', 'destination_ID','distances', 'volume_in_unipartite','dest_masses','results_all=False']]
-        
-        flow['acc_uni'] = Accessibility(flow_df = flow, all_destinations=False, function = AFED)
+        flow['acc_uni'] = function(flow_df = flow, all_destinations=False)
         
         self.assertEqual(flow['results_all=False'].all(), flow['acc_uni'].all())
